copied over all tests
