+++ tensorzero-core/src/inference/types/streams.rs
@@ -563,8 +563,19 @@ fn handle_textual_content_block<F, A>(
 
 #[cfg(test)]
 mod tests {
+    use std::collections::{HashMap, HashSet};
+
     use super::*;
-    use crate::inference::types::{ContentBlockOutputType, Text, Thought};
+    use crate::{
+        config::SchemaData,
+        function::{FunctionConfigChat, FunctionConfigJson},
+        inference::types::{
+            current_timestamp, ContentBlockChatOutput, ContentBlockOutputType, InferenceResult,
+            Text, Thought,
+        },
+        jsonschema_util::StaticJSONSchema,
+        tool::ToolCallOutput,
+    };
 
     #[test]
     fn test_handle_textual_content_block() {
@@ -683,5 +694,1433 @@ mod tests {
             }
             _ => panic!("Expected thought block"),
         }
+    }
+    #[tokio::test]
+    async fn test_collect_chunks() {
+        // Test case 1: empty chunks (should error)
+        let templates = TemplateConfig::default();
+
+        let chunks = vec![];
+        let function_config = Arc::new(FunctionConfig::Chat(FunctionConfigChat::default()));
+        let model_name = "test_model";
+        let model_provider_name = "test_provider";
+        let raw_request = "raw request".to_string();
+        let collect_chunks_args = CollectChunksArgs {
+            inference_id: Uuid::now_v7(),
+            episode_id: Uuid::now_v7(),
+            value: chunks,
+            system: None,
+            input_messages: vec![],
+            function: function_config.clone(),
+            model_name: model_name.into(),
+            model_provider_name: model_provider_name.into(),
+            raw_request: raw_request.clone(),
+            raw_response: None,
+            inference_params: InferenceParams::default(),
+            function_name: "",
+            variant_name: "",
+            dynamic_output_schema: None,
+            templates: &templates,
+            tool_config: None,
+            cached: false,
+            extra_body: Default::default(),
+            extra_headers: Default::default(),
+            fetch_and_encode_input_files_before_inference: false,
+        };
+        let result = collect_chunks(collect_chunks_args).await;
+        assert_eq!(
+            result.unwrap_err(),
+            ErrorDetails::TypeConversion {
+                message:
+                    "Attempted to create an InferenceResult from an empty response chunk vector"
+                        .to_string(),
+            }
+            .into()
+        );
+
+        // Test case 2: non-empty chunks with no tool calls but content exists
+        let inference_id = Uuid::now_v7();
+        let episode_id = Uuid::now_v7();
+        let created = current_timestamp();
+        let content = vec![ContentBlockChunk::Text(TextChunk {
+            text: "Hello,".to_string(),
+            id: "0".to_string(),
+        })];
+        let latency = Duration::from_millis(150);
+        let chunks = vec![
+            InferenceResultChunk::Chat(ChatInferenceResultChunk {
+                content,
+                created,
+                usage: None,
+                raw_response: "{\"message\": \"Hello}".to_string(),
+                latency,
+                finish_reason: None,
+            }),
+            InferenceResultChunk::Chat(ChatInferenceResultChunk {
+                content: vec![ContentBlockChunk::Text(TextChunk {
+                    text: " world!".to_string(),
+                    id: "0".to_string(),
+                })],
+                created,
+                usage: Some(Usage {
+                    input_tokens: 2,
+                    output_tokens: 4,
+                }),
+                raw_response: ", world!\"}".to_string(),
+                latency: Duration::from_millis(250),
+                finish_reason: Some(FinishReason::Stop),
+            }),
+        ];
+        let collect_chunks_args = CollectChunksArgs {
+            inference_id,
+            episode_id,
+            value: chunks,
+            system: None,
+            input_messages: vec![],
+            function: function_config.clone(),
+            model_name: model_name.into(),
+            model_provider_name: model_provider_name.into(),
+            raw_request: raw_request.clone(),
+            raw_response: None,
+            inference_params: InferenceParams::default(),
+            function_name: "",
+            variant_name: "",
+            dynamic_output_schema: None,
+            templates: &templates,
+            tool_config: None,
+            cached: false,
+            extra_body: Default::default(),
+            extra_headers: Default::default(),
+            fetch_and_encode_input_files_before_inference: false,
+        };
+        let result = collect_chunks(collect_chunks_args).await.unwrap();
+        let chat_result = match result {
+            InferenceResult::Chat(chat_result) => chat_result,
+            InferenceResult::Json(_) => panic!("Expected Chat inference response"),
+        };
+        assert_eq!(chat_result.inference_id, inference_id);
+        // We make a new timestamp for `chat_result.created`, so just check that it's at least
+        // the timestamp of the first chunk.
+        assert!(
+            chat_result.created >= created,
+            "Chat result was created at {:?}, before the first chunk was created at {:?}",
+            chat_result.created,
+            created
+        );
+        assert_eq!(chat_result.finish_reason, Some(FinishReason::Stop));
+        assert_eq!(
+            chat_result.content,
+            vec!["Hello, world!".to_string().into()]
+        );
+        assert_eq!(chat_result.model_inference_results.len(), 1);
+        let model_inference_result = chat_result.model_inference_results.first().unwrap();
+        assert_eq!(&*model_inference_result.model_name, model_name);
+        assert_eq!(
+            &*model_inference_result.model_provider_name,
+            model_provider_name
+        );
+        assert_eq!(model_inference_result.raw_request, raw_request);
+        // Test Case 3: a JSON string that passes validation and also include usage in each chunk
+        let inference_id = Uuid::now_v7();
+        let output_schema = serde_json::json!({
+            "type": "object",
+            "properties": {
+                "name": {"type": "string"},
+                "age": {"type": "number"}
+            },
+            "required": ["name", "age"]
+        });
+        let implicit_tool_call_config = ToolCallConfig::implicit_from_value(&output_schema);
+        let output_schema = StaticJSONSchema::from_value(output_schema).unwrap();
+        let json_function_config = Arc::new(FunctionConfig::Json(FunctionConfigJson {
+            variants: HashMap::new(),
+            schemas: SchemaData::default(),
+            implicit_tool_call_config,
+            output_schema,
+            description: None,
+            all_explicit_template_names: HashSet::new(),
+        }));
+        let usage1 = Usage {
+            input_tokens: 10,
+            output_tokens: 5,
+        };
+        let usage2 = Usage {
+            input_tokens: 5,
+            output_tokens: 10,
+        };
+        let chunks = vec![
+            InferenceResultChunk::Json(JsonInferenceResultChunk {
+                raw: Some("{\"name\":".to_string()),
+                thought: Some("Thought 1".to_string()),
+                created,
+                usage: Some(usage1),
+                raw_response: "{\"name\":".to_string(),
+                latency: Duration::from_millis(150),
+                finish_reason: Some(FinishReason::ToolCall),
+            }),
+            InferenceResultChunk::Json(JsonInferenceResultChunk {
+                raw: Some("\"John\",\"age\":30}".to_string()),
+                thought: Some("Thought 2".to_string()),
+                created,
+                usage: Some(usage2),
+                raw_response: "\"John\",\"age\":30}".to_string(),
+                latency: Duration::from_millis(250),
+                finish_reason: Some(FinishReason::Stop),
+            }),
+        ];
+        let collect_chunks_args = CollectChunksArgs {
+            value: chunks,
+            system: None,
+            inference_id,
+            episode_id,
+            input_messages: vec![],
+            function: json_function_config.clone(),
+            model_name: model_name.into(),
+            model_provider_name: model_provider_name.into(),
+            raw_request: raw_request.clone(),
+            raw_response: None,
+            inference_params: InferenceParams::default(),
+            function_name: "",
+            variant_name: "",
+            dynamic_output_schema: None,
+            templates: &templates,
+            tool_config: None,
+            cached: false,
+            extra_body: Default::default(),
+            extra_headers: Default::default(),
+            fetch_and_encode_input_files_before_inference: false,
+        };
+        let response = collect_chunks(collect_chunks_args).await.unwrap();
+        assert_eq!(
+            response.usage_considering_cached(),
+            Usage {
+                input_tokens: 15,
+                output_tokens: 15,
+            }
+        );
+        match response {
+            InferenceResult::Json(json_result) => {
+                assert_eq!(json_result.inference_id, inference_id);
+                assert_eq!(
+                    json_result.output.parsed,
+                    Some(serde_json::json!({"name": "John", "age": 30}))
+                );
+                assert_eq!(
+                    json_result.output.raw,
+                    Some("{\"name\":\"John\",\"age\":30}".to_string())
+                );
+                assert_eq!(json_result.finish_reason, Some(FinishReason::Stop));
+                assert_eq!(json_result.model_inference_results.len(), 1);
+                let model_inference_result = json_result.model_inference_results.first().unwrap();
+                assert_eq!(&*model_inference_result.model_name, model_name);
+                assert_eq!(
+                    &*model_inference_result.model_provider_name,
+                    model_provider_name
+                );
+                assert_eq!(model_inference_result.raw_request, raw_request);
+            }
+            InferenceResult::Chat(_) => panic!("Expected Json inference response"),
+        }
+
+        // Test Case 4: a JSON string that fails validation and usage only in last chunk
+        let inference_id = Uuid::now_v7();
+        let created = current_timestamp();
+        let usage = Usage {
+            input_tokens: 10,
+            output_tokens: 5,
+        };
+        let chunks = vec![
+            InferenceResultChunk::Json(JsonInferenceResultChunk {
+                raw: Some("{\"name\":".to_string()),
+                thought: Some("Thought 1".to_string()),
+                created,
+                usage: Some(usage),
+                raw_response: "{\"name\":".to_string(),
+                latency: Duration::from_millis(100),
+                finish_reason: Some(FinishReason::ToolCall),
+            }),
+            InferenceResultChunk::Json(JsonInferenceResultChunk {
+                raw: Some("\"John\"}".to_string()),
+                thought: None,
+                created,
+                usage: None,
+                raw_response: "\"John\"}".to_string(),
+                latency: Duration::from_millis(200),
+                finish_reason: None,
+            }),
+        ];
+        let collect_chunks_args = CollectChunksArgs {
+            value: chunks,
+            inference_id,
+            episode_id,
+            system: None,
+            input_messages: vec![],
+            function: json_function_config.clone(),
+            model_name: model_name.into(),
+            model_provider_name: model_provider_name.into(),
+            raw_request: raw_request.clone(),
+            raw_response: None,
+            inference_params: InferenceParams::default(),
+            function_name: "",
+            variant_name: "",
+            dynamic_output_schema: None,
+            templates: &templates,
+            tool_config: None,
+            cached: false,
+            extra_body: Default::default(),
+            extra_headers: Default::default(),
+            fetch_and_encode_input_files_before_inference: false,
+        };
+        let result = collect_chunks(collect_chunks_args).await.unwrap();
+        assert_eq!(result.usage_considering_cached(), usage);
+        match result {
+            InferenceResult::Json(json_result) => {
+                assert_eq!(json_result.inference_id, inference_id);
+                // We make a new timestamp for `json_result.created`, so just check that it's at least
+                // the timestamp of the first chunk.
+                assert!(
+                    json_result.created >= created,
+                    "Json result was created at {:?}, before the first chunk was created at {:?}",
+                    json_result.created,
+                    created
+                );
+                assert_eq!(json_result.output.parsed, None);
+                assert_eq!(
+                    json_result.output.raw,
+                    Some("{\"name\":\"John\"}".to_string())
+                );
+                assert_eq!(json_result.model_inference_results.len(), 1);
+                let model_inference_result = json_result.model_inference_results.first().unwrap();
+                assert_eq!(&*model_inference_result.model_name, model_name);
+                assert_eq!(
+                    &*model_inference_result.model_provider_name,
+                    model_provider_name
+                );
+                assert_eq!(model_inference_result.raw_request, raw_request);
+            }
+            InferenceResult::Chat(_) => panic!("Expected Json inference response"),
+        }
+
+        // Test case 5: chunks with some None content
+        let inference_id = Uuid::now_v7();
+        let episode_id = Uuid::now_v7();
+        let created = current_timestamp();
+        let usage = Usage {
+            input_tokens: 15,
+            output_tokens: 10,
+        };
+        let chunks = vec![
+            InferenceResultChunk::Json(JsonInferenceResultChunk {
+                raw: Some("{\"name\":\"John\",".to_string()),
+                thought: None,
+                created,
+                usage: Some(usage),
+                raw_response: "{\"name\":\"John\",".to_string(),
+                latency: Duration::from_millis(100),
+                finish_reason: None,
+            }),
+            InferenceResultChunk::Json(JsonInferenceResultChunk {
+                raw: Some(String::new()),
+                thought: Some("Thought 2".to_string()),
+                created,
+                usage: None,
+                raw_response: String::new(),
+                latency: Duration::from_millis(200),
+                finish_reason: None,
+            }),
+            InferenceResultChunk::Json(JsonInferenceResultChunk {
+                raw: Some("\"age\":30}".to_string()),
+                thought: None,
+                created,
+                usage: None,
+                raw_response: "\"age\":30}".to_string(),
+                latency: Duration::from_millis(300),
+                finish_reason: Some(FinishReason::Stop),
+            }),
+        ];
+        let collect_chunks_args = CollectChunksArgs {
+            value: chunks,
+            inference_id,
+            episode_id,
+            system: None,
+            input_messages: vec![],
+            function: function_config.clone(),
+            model_name: model_name.into(),
+            model_provider_name: model_provider_name.into(),
+            raw_request: raw_request.clone(),
+            raw_response: None,
+            inference_params: InferenceParams::default(),
+            function_name: "",
+            variant_name: "",
+            dynamic_output_schema: None,
+            templates: &templates,
+            tool_config: None,
+            cached: false,
+            extra_body: Default::default(),
+            extra_headers: Default::default(),
+            fetch_and_encode_input_files_before_inference: false,
+        };
+        let result = collect_chunks(collect_chunks_args).await.unwrap();
+        assert_eq!(result.usage_considering_cached(), usage);
+        match result {
+            InferenceResult::Chat(chat_response) => {
+                assert_eq!(chat_response.inference_id, inference_id);
+                // We make a new timestamp for `chat_response.created`, so just check that it's at least
+                // the timestamp of the first chunk.
+                assert!(
+                    chat_response.created >= created,
+                    "Chat result was created at {:?}, before the first chunk was created at {:?}",
+                    chat_response.created,
+                    created
+                );
+                assert_eq!(
+                    chat_response.content,
+                    vec![
+                        ContentBlockChatOutput::Text(Text {
+                            text: "{\"name\":\"John\",\"age\":30}".to_string()
+                        }),
+                        ContentBlockChatOutput::Thought(Thought {
+                            text: Some("Thought 2".to_string()),
+                            signature: None,
+                            provider_type: None,
+                        }),
+                    ]
+                );
+                assert_eq!(chat_response.model_inference_results.len(), 1);
+                let model_inference_result = chat_response.model_inference_results.first().unwrap();
+                assert_eq!(&*model_inference_result.model_name, model_name);
+                assert_eq!(chat_response.finish_reason, Some(FinishReason::Stop));
+                assert_eq!(
+                    model_inference_result.finish_reason,
+                    Some(FinishReason::Stop)
+                );
+                assert_eq!(
+                    &*model_inference_result.model_provider_name,
+                    model_provider_name
+                );
+                assert_eq!(model_inference_result.raw_request, raw_request);
+            }
+            InferenceResult::Json(_) => panic!("Expected Chat inference response"),
+        }
+
+        // Test Case 6: a JSON function with implicit tool call config
+        let inference_id = Uuid::now_v7();
+        let created = current_timestamp();
+        let output_schema = serde_json::json!({
+            "type": "object",
+            "properties": {
+                "name": {"type": "string"},
+                "age": {"type": "number"}
+            },
+            "required": ["name", "age"]
+        });
+        let implicit_tool_call_config = ToolCallConfig::implicit_from_value(&output_schema);
+        let output_schema = StaticJSONSchema::from_value(output_schema).unwrap();
+        let json_function_config = Arc::new(FunctionConfig::Json(FunctionConfigJson {
+            variants: HashMap::new(),
+            schemas: SchemaData::default(),
+            implicit_tool_call_config,
+            output_schema,
+            description: None,
+            all_explicit_template_names: HashSet::new(),
+        }));
+        let usage1 = Usage {
+            input_tokens: 10,
+            output_tokens: 5,
+        };
+        let usage2 = Usage {
+            input_tokens: 5,
+            output_tokens: 10,
+        };
+        let chunks = vec![
+            InferenceResultChunk::Json(JsonInferenceResultChunk {
+                raw: Some("{\"name\":".to_string()),
+                thought: Some("Thought 1".to_string()),
+                created,
+                usage: Some(usage1),
+                raw_response: "{\"name\":".to_string(),
+                latency: Duration::from_millis(150),
+                finish_reason: Some(FinishReason::ToolCall),
+            }),
+            InferenceResultChunk::Json(JsonInferenceResultChunk {
+                raw: Some("\"John\",\"age\":30}".to_string()),
+                thought: Some("Thought 2".to_string()),
+                created,
+                usage: Some(usage2),
+                raw_response: "\"John\",\"age\":30}".to_string(),
+                latency: Duration::from_millis(250),
+                finish_reason: Some(FinishReason::Stop),
+            }),
+        ];
+        let collect_chunks_args = CollectChunksArgs {
+            inference_id,
+            episode_id,
+            value: chunks,
+            system: None,
+            input_messages: vec![],
+            function: json_function_config.clone(),
+            model_name: model_name.into(),
+            model_provider_name: model_provider_name.into(),
+            raw_request: raw_request.clone(),
+            raw_response: None,
+            inference_params: InferenceParams::default(),
+            function_name: "",
+            variant_name: "",
+            dynamic_output_schema: None,
+            templates: &templates,
+            tool_config: None,
+            cached: false,
+            extra_body: Default::default(),
+            extra_headers: Default::default(),
+            fetch_and_encode_input_files_before_inference: false,
+        };
+        let response = collect_chunks(collect_chunks_args).await.unwrap();
+        assert_eq!(
+            response.usage_considering_cached(),
+            Usage {
+                input_tokens: 15,
+                output_tokens: 15,
+            }
+        );
+        match response {
+            InferenceResult::Json(json_result) => {
+                assert_eq!(json_result.inference_id, inference_id);
+                assert_eq!(
+                    json_result.output.parsed,
+                    Some(serde_json::json!({"name": "John", "age": 30}))
+                );
+                assert_eq!(
+                    json_result.output.raw,
+                    Some("{\"name\":\"John\",\"age\":30}".to_string())
+                );
+                assert_eq!(json_result.finish_reason, Some(FinishReason::Stop));
+                assert_eq!(json_result.model_inference_results.len(), 1);
+                let model_inference_result = json_result.model_inference_results.first().unwrap();
+                assert_eq!(&*model_inference_result.model_name, model_name);
+                assert_eq!(
+                    &*model_inference_result.model_provider_name,
+                    model_provider_name
+                );
+                assert_eq!(model_inference_result.raw_request, raw_request);
+            }
+            InferenceResult::Chat(_) => panic!("Expected Json inference response"),
+        }
+        // Test Case 7: a JSON string with a dynamic schema that passes validation and also include usage in each chunk
+        let inference_id = Uuid::now_v7();
+        let static_output_schema = serde_json::json!({
+            "type": "object",
+            "properties": {
+                "name": {"type": "string"},
+            },
+            "required": ["name"]
+        });
+        let implicit_tool_call_config = ToolCallConfig::implicit_from_value(&static_output_schema);
+        let output_schema = StaticJSONSchema::from_value(static_output_schema).unwrap();
+        let json_function_config = Arc::new(FunctionConfig::Json(FunctionConfigJson {
+            variants: HashMap::new(),
+            schemas: SchemaData::default(),
+            implicit_tool_call_config,
+            output_schema,
+            description: None,
+            all_explicit_template_names: HashSet::new(),
+        }));
+        let usage1 = Usage {
+            input_tokens: 10,
+            output_tokens: 5,
+        };
+        let usage2 = Usage {
+            input_tokens: 5,
+            output_tokens: 10,
+        };
+        let dynamic_output_schema = DynamicJSONSchema::new(serde_json::json!({
+            "type": "object",
+            "properties": {
+                "name": {"type": "string"},
+                "age": {"type": "number"}
+            },
+            "required": ["name", "age"]
+        }));
+        let templates = TemplateConfig::default();
+        let chunks = vec![
+            InferenceResultChunk::Json(JsonInferenceResultChunk {
+                raw: Some("{\"name\":".to_string()),
+                thought: Some("Thought 1".to_string()),
+                created,
+                usage: Some(usage1),
+                finish_reason: Some(FinishReason::Stop),
+                raw_response: "{\"name\":".to_string(),
+                latency: Duration::from_millis(150),
+            }),
+            InferenceResultChunk::Json(JsonInferenceResultChunk {
+                raw: Some("\"John\",\"age\":30}".to_string()),
+                thought: Some("Thought 2".to_string()),
+                created,
+                usage: Some(usage2),
+                finish_reason: Some(FinishReason::ToolCall),
+                raw_response: "\"John\",\"age\":30}".to_string(),
+                latency: Duration::from_millis(250),
+            }),
+        ];
+        let collect_chunks_args = CollectChunksArgs {
+            inference_id,
+            episode_id,
+            value: chunks,
+            system: None,
+            input_messages: vec![],
+            function: json_function_config.clone(),
+            model_name: model_name.into(),
+            model_provider_name: model_provider_name.into(),
+            raw_request: raw_request.clone(),
+            raw_response: None,
+            inference_params: InferenceParams::default(),
+            function_name: "",
+            variant_name: "",
+            dynamic_output_schema: Some(dynamic_output_schema),
+            templates: &templates,
+            tool_config: None,
+            cached: false,
+            extra_body: Default::default(),
+            extra_headers: Default::default(),
+            fetch_and_encode_input_files_before_inference: false,
+        };
+        let response = collect_chunks(collect_chunks_args).await.unwrap();
+        assert_eq!(
+            response.usage_considering_cached(),
+            Usage {
+                input_tokens: 15,
+                output_tokens: 15,
+            }
+        );
+        match response {
+            InferenceResult::Json(json_result) => {
+                assert_eq!(json_result.inference_id, inference_id);
+                assert_eq!(
+                    json_result.output.parsed,
+                    Some(serde_json::json!({"name": "John", "age": 30}))
+                );
+                assert_eq!(
+                    json_result.output.raw,
+                    Some("{\"name\":\"John\",\"age\":30}".to_string())
+                );
+                assert_eq!(json_result.model_inference_results.len(), 1);
+                let model_inference_result = json_result.model_inference_results.first().unwrap();
+                assert_eq!(&*model_inference_result.model_name, model_name);
+                assert_eq!(
+                    model_inference_result.finish_reason,
+                    Some(FinishReason::ToolCall)
+                );
+                assert_eq!(
+                    &*model_inference_result.model_provider_name,
+                    model_provider_name
+                );
+                assert_eq!(model_inference_result.raw_request, raw_request);
+            }
+            InferenceResult::Chat(_) => panic!("Expected Json inference response"),
+        }
+    }
+
+    #[tokio::test]
+    async fn test_collect_interleaved_chunks() {
+        let templates = TemplateConfig::default();
+        let function_config = Arc::new(FunctionConfig::Chat(FunctionConfigChat::default()));
+        let model_name = "test_model";
+        let model_provider_name = "test_provider";
+        let raw_request = "raw request".to_string();
+        let inference_id = Uuid::now_v7();
+        let episode_id = Uuid::now_v7();
+        let created = current_timestamp();
+        let latency = Duration::from_millis(150);
+        let chunks = vec![
+            InferenceResultChunk::Chat(ChatInferenceResultChunk {
+                content: vec![
+                    ContentBlockChunk::Text(TextChunk {
+                        text: "Hello ".to_string(),
+                        id: "0".to_string(),
+                    }),
+                    ContentBlockChunk::ToolCall(ToolCallChunk {
+                        id: "0".to_string(),
+                        raw_name: Some("my_tool_call".to_string()),
+                        raw_arguments: "true".to_string(),
+                    }),
+                ],
+                created,
+                usage: None,
+                raw_response: "{\"message\": \"Hello}".to_string(),
+                latency,
+                finish_reason: None,
+            }),
+            InferenceResultChunk::Chat(ChatInferenceResultChunk {
+                content: vec![
+                    ContentBlockChunk::Thought(ThoughtChunk {
+                        text: Some("Some thou".to_string()),
+                        id: "0".to_string(),
+                        signature: None,
+                        provider_type: None,
+                    }),
+                    ContentBlockChunk::Thought(ThoughtChunk {
+                        text: Some("My other interleaved thought".to_string()),
+                        id: "1".to_string(),
+                        signature: None,
+                        provider_type: None,
+                    }),
+                ],
+                created,
+                usage: None,
+                raw_response: "my raw thought".to_string(),
+                latency,
+                finish_reason: None,
+            }),
+            InferenceResultChunk::Chat(ChatInferenceResultChunk {
+                content: vec![ContentBlockChunk::Text(TextChunk {
+                    text: "world!".to_string(),
+                    id: "0".to_string(),
+                })],
+                created,
+                usage: Some(Usage {
+                    input_tokens: 2,
+                    output_tokens: 4,
+                }),
+                raw_response: ", world!\"}".to_string(),
+                latency: Duration::from_millis(250),
+                finish_reason: Some(FinishReason::Stop),
+            }),
+            InferenceResultChunk::Chat(ChatInferenceResultChunk {
+                content: vec![ContentBlockChunk::Thought(ThoughtChunk {
+                    text: Some("ght".to_string()),
+                    id: "0".to_string(),
+                    signature: None,
+                    provider_type: None,
+                })],
+                created,
+                usage: None,
+                raw_response: "my other raw thought".to_string(),
+                latency,
+                finish_reason: None,
+            }),
+        ];
+        let collect_chunks_args = CollectChunksArgs {
+            inference_id,
+            episode_id,
+            value: chunks,
+            system: None,
+            input_messages: vec![],
+            function: function_config.clone(),
+            model_name: model_name.into(),
+            model_provider_name: model_provider_name.into(),
+            raw_request: raw_request.clone(),
+            raw_response: None,
+            inference_params: InferenceParams::default(),
+            function_name: "",
+            variant_name: "",
+            dynamic_output_schema: None,
+            templates: &templates,
+            tool_config: None,
+            cached: false,
+            extra_body: Default::default(),
+            extra_headers: Default::default(),
+            fetch_and_encode_input_files_before_inference: false,
+        };
+        let result = collect_chunks(collect_chunks_args).await.unwrap();
+        assert_eq!(
+            result.usage_considering_cached(),
+            Usage {
+                input_tokens: 2,
+                output_tokens: 4,
+            }
+        );
+        let chat_result = match result {
+            InferenceResult::Chat(chat_result) => chat_result,
+            InferenceResult::Json(_) => panic!("Expected Chat inference response"),
+        };
+        assert_eq!(chat_result.inference_id, inference_id);
+        // We make a new timestamp for `chat_result.created`, so just check that it's at least
+        // the timestamp of the first chunk.
+        assert!(
+            chat_result.created >= created,
+            "Chat result was created at {:?}, before the first chunk was created at {:?}",
+            chat_result.created,
+            created
+        );
+        assert_eq!(chat_result.finish_reason, Some(FinishReason::Stop));
+
+        let expected_content = vec![
+            ContentBlockChatOutput::Text(Text {
+                text: "Hello world!".to_string(),
+            }),
+            ContentBlockChatOutput::ToolCall(ToolCallOutput {
+                name: None,
+                raw_name: "my_tool_call".to_string(),
+                raw_arguments: "true".to_string(),
+                arguments: None,
+                id: "0".to_string(),
+            }),
+            ContentBlockChatOutput::Thought(Thought {
+                text: Some("Some thought".to_string()),
+                signature: None,
+                provider_type: None,
+            }),
+            ContentBlockChatOutput::Thought(Thought {
+                text: Some("My other interleaved thought".to_string()),
+                signature: None,
+                provider_type: None,
+            }),
+        ];
+        assert_eq!(chat_result.content, expected_content);
+
+        assert_eq!(chat_result.model_inference_results.len(), 1);
+        let model_inference_result = chat_result.model_inference_results.first().unwrap();
+        assert_eq!(&*model_inference_result.model_name, model_name);
+        assert_eq!(
+            &*model_inference_result.model_provider_name,
+            model_provider_name
+        );
+        assert_eq!(model_inference_result.raw_request, raw_request);
+    }
+
+    #[tokio::test]
+    async fn test_collect_chunks_tool_name_accumulation() {
+        let templates = TemplateConfig::default();
+        let function_config = Arc::new(FunctionConfig::Chat(FunctionConfigChat::default()));
+        let model_name = "test_model";
+        let model_provider_name = "test_provider";
+        let raw_request = "raw request".to_string();
+        let inference_id = Uuid::now_v7();
+        let episode_id = Uuid::now_v7();
+        let created = current_timestamp();
+        let latency = Duration::from_millis(150);
+
+        // Test case 1: Tool name sent in first chunk, then arguments accumulated
+        let chunks_case1 = vec![
+            InferenceResultChunk::Chat(ChatInferenceResultChunk {
+                content: vec![ContentBlockChunk::ToolCall(ToolCallChunk {
+                    id: "tool_1".to_string(),
+                    raw_name: Some("get_weather".to_string()),
+                    raw_arguments: "{\"loca".to_string(),
+                })],
+                created,
+                usage: None,
+                raw_response: "chunk1".to_string(),
+                latency,
+                finish_reason: None,
+            }),
+            InferenceResultChunk::Chat(ChatInferenceResultChunk {
+                content: vec![ContentBlockChunk::ToolCall(ToolCallChunk {
+                    id: "tool_1".to_string(),
+                    raw_name: None, // No name in subsequent chunks
+                    raw_arguments: "tion\": \"San Francisco\", \"unit\": \"celsius\"}".to_string(),
+                })],
+                created,
+                usage: Some(Usage {
+                    input_tokens: 10,
+                    output_tokens: 20,
+                }),
+                raw_response: "chunk2".to_string(),
+                latency: Duration::from_millis(250),
+                finish_reason: Some(FinishReason::ToolCall),
+            }),
+        ];
+
+        let collect_chunks_args = CollectChunksArgs {
+            inference_id,
+            episode_id,
+            value: chunks_case1,
+            system: None,
+            input_messages: vec![],
+            function: function_config.clone(),
+            model_name: model_name.into(),
+            model_provider_name: model_provider_name.into(),
+            raw_request: raw_request.clone(),
+            raw_response: None,
+            inference_params: InferenceParams::default(),
+            function_name: "",
+            variant_name: "",
+            dynamic_output_schema: None,
+            templates: &templates,
+            tool_config: None,
+            cached: false,
+            extra_body: Default::default(),
+            extra_headers: Default::default(),
+            fetch_and_encode_input_files_before_inference: false,
+        };
+
+        let result = collect_chunks(collect_chunks_args).await.unwrap();
+        let chat_result = match result {
+            InferenceResult::Chat(chat_result) => chat_result,
+            InferenceResult::Json(_) => panic!("Expected Chat inference response"),
+        };
+
+        assert_eq!(chat_result.content.len(), 1);
+        match &chat_result.content[0] {
+            ContentBlockChatOutput::ToolCall(tool_call) => {
+                assert_eq!(tool_call.raw_name, "get_weather");
+                assert_eq!(
+                    tool_call.raw_arguments,
+                    r#"{"location": "San Francisco", "unit": "celsius"}"#
+                );
+                assert_eq!(tool_call.id, "tool_1");
+            }
+            _ => panic!("Expected tool call block"),
+        }
+
+        // Test case 2: Multiple tool calls with different IDs and name accumulation
+        let chunks_case2 = vec![
+            InferenceResultChunk::Chat(ChatInferenceResultChunk {
+                content: vec![
+                    ContentBlockChunk::ToolCall(ToolCallChunk {
+                        id: "tool_1".to_string(),
+                        raw_name: Some("get_wea".to_string()),
+                        raw_arguments: "{\"loc".to_string(),
+                    }),
+                    ContentBlockChunk::ToolCall(ToolCallChunk {
+                        id: "tool_2".to_string(),
+                        raw_name: Some("calculate".to_string()),
+                        raw_arguments: "{\"expr".to_string(),
+                    }),
+                ],
+                created,
+                usage: None,
+                raw_response: "chunk1".to_string(),
+                latency,
+                finish_reason: None,
+            }),
+            InferenceResultChunk::Chat(ChatInferenceResultChunk {
+                content: vec![
+                    ContentBlockChunk::ToolCall(ToolCallChunk {
+                        id: "tool_1".to_string(),
+                        raw_name: Some("ther".to_string()), // Continue accumulating name
+                        raw_arguments: "ation\": \"NYC\"}".to_string(),
+                    }),
+                    ContentBlockChunk::ToolCall(ToolCallChunk {
+                        id: "tool_2".to_string(),
+                        raw_name: None, // No more name for tool_2
+                        raw_arguments: "ession\": \"2+2\"}".to_string(),
+                    }),
+                ],
+                created,
+                usage: Some(Usage {
+                    input_tokens: 15,
+                    output_tokens: 25,
+                }),
+                raw_response: "chunk2".to_string(),
+                latency: Duration::from_millis(250),
+                finish_reason: Some(FinishReason::ToolCall),
+            }),
+        ];
+
+        let collect_chunks_args = CollectChunksArgs {
+            inference_id,
+            episode_id,
+            value: chunks_case2,
+            system: None,
+            input_messages: vec![],
+            function: function_config.clone(),
+            model_name: model_name.into(),
+            model_provider_name: model_provider_name.into(),
+            raw_request: raw_request.clone(),
+            raw_response: None,
+            inference_params: InferenceParams::default(),
+            function_name: "",
+            variant_name: "",
+            dynamic_output_schema: None,
+            templates: &templates,
+            tool_config: None,
+            cached: false,
+            extra_body: Default::default(),
+            extra_headers: Default::default(),
+            fetch_and_encode_input_files_before_inference: false,
+        };
+
+        let result = collect_chunks(collect_chunks_args).await.unwrap();
+        let chat_result = match result {
+            InferenceResult::Chat(chat_result) => chat_result,
+            InferenceResult::Json(_) => panic!("Expected Chat inference response"),
+        };
+
+        assert_eq!(chat_result.content.len(), 2);
+        match &chat_result.content[0] {
+            ContentBlockChatOutput::ToolCall(tool_call) => {
+                assert_eq!(tool_call.raw_name, "get_weather");
+                assert_eq!(tool_call.raw_arguments, r#"{"location": "NYC"}"#);
+                assert_eq!(tool_call.id, "tool_1");
+            }
+            _ => panic!("Expected first tool call block"),
+        }
+        match &chat_result.content[1] {
+            ContentBlockChatOutput::ToolCall(tool_call) => {
+                assert_eq!(tool_call.raw_name, "calculate");
+                assert_eq!(tool_call.raw_arguments, r#"{"expression": "2+2"}"#);
+                assert_eq!(tool_call.id, "tool_2");
+            }
+            _ => panic!("Expected second tool call block"),
+        }
+
+        // Test case 3: Tool call with no name in first chunk (should start with empty name)
+        let chunks_case3 = vec![
+            InferenceResultChunk::Chat(ChatInferenceResultChunk {
+                content: vec![ContentBlockChunk::ToolCall(ToolCallChunk {
+                    id: "tool_1".to_string(),
+                    raw_name: None, // No name in first chunk
+                    raw_arguments: "{\"key\":".to_string(),
+                })],
+                created,
+                usage: None,
+                raw_response: "chunk1".to_string(),
+                latency,
+                finish_reason: None,
+            }),
+            InferenceResultChunk::Chat(ChatInferenceResultChunk {
+                content: vec![ContentBlockChunk::ToolCall(ToolCallChunk {
+                    id: "tool_1".to_string(),
+                    raw_name: Some("my_function".to_string()), // Name comes later
+                    raw_arguments: " \"value\"}".to_string(),
+                })],
+                created,
+                usage: Some(Usage {
+                    input_tokens: 5,
+                    output_tokens: 10,
+                }),
+                raw_response: "chunk2".to_string(),
+                latency: Duration::from_millis(250),
+                finish_reason: Some(FinishReason::ToolCall),
+            }),
+        ];
+
+        let collect_chunks_args = CollectChunksArgs {
+            inference_id,
+            episode_id,
+            value: chunks_case3,
+            system: None,
+            input_messages: vec![],
+            function: function_config.clone(),
+            model_name: model_name.into(),
+            model_provider_name: model_provider_name.into(),
+            raw_request: raw_request.clone(),
+            raw_response: None,
+            inference_params: InferenceParams::default(),
+            function_name: "",
+            variant_name: "",
+            dynamic_output_schema: None,
+            templates: &templates,
+            tool_config: None,
+            cached: false,
+            extra_body: Default::default(),
+            extra_headers: Default::default(),
+            fetch_and_encode_input_files_before_inference: false,
+        };
+
+        let result = collect_chunks(collect_chunks_args).await.unwrap();
+        let chat_result = match result {
+            InferenceResult::Chat(chat_result) => chat_result,
+            InferenceResult::Json(_) => panic!("Expected Chat inference response"),
+        };
+
+        assert_eq!(chat_result.content.len(), 1);
+        match &chat_result.content[0] {
+            ContentBlockChatOutput::ToolCall(tool_call) => {
+                assert_eq!(tool_call.raw_name, "my_function"); // Should accumulate to the full name
+                assert_eq!(tool_call.raw_arguments, r#"{"key": "value"}"#);
+                assert_eq!(tool_call.id, "tool_1");
+            }
+            _ => panic!("Expected tool call block"),
+        }
+
+        // Test case 4: Mixed content with text and tool calls preserving order
+        let chunks_case4 = vec![
+            InferenceResultChunk::Chat(ChatInferenceResultChunk {
+                content: vec![
+                    ContentBlockChunk::Text(TextChunk {
+                        text: "I'll help you with that. ".to_string(),
+                        id: "0".to_string(),
+                    }),
+                    ContentBlockChunk::ToolCall(ToolCallChunk {
+                        id: "tool_1".to_string(),
+                        raw_name: Some("search".to_string()),
+                        raw_arguments: "{\"query\"".to_string(),
+                    }),
+                ],
+                created,
+                usage: None,
+                raw_response: "chunk1".to_string(),
+                latency,
+                finish_reason: None,
+            }),
+            InferenceResultChunk::Chat(ChatInferenceResultChunk {
+                content: vec![
+                    ContentBlockChunk::Text(TextChunk {
+                        text: "Let me search for information.".to_string(),
+                        id: "0".to_string(),
+                    }),
+                    ContentBlockChunk::ToolCall(ToolCallChunk {
+                        id: "tool_1".to_string(),
+                        raw_name: None,
+                        raw_arguments: ": \"weather today\"}".to_string(),
+                    }),
+                ],
+                created,
+                usage: Some(Usage {
+                    input_tokens: 20,
+                    output_tokens: 15,
+                }),
+                raw_response: "chunk2".to_string(),
+                latency: Duration::from_millis(250),
+                finish_reason: Some(FinishReason::ToolCall),
+            }),
+        ];
+
+        let collect_chunks_args = CollectChunksArgs {
+            inference_id,
+            episode_id,
+            value: chunks_case4,
+            system: None,
+            input_messages: vec![],
+            function: function_config.clone(),
+            model_name: model_name.into(),
+            model_provider_name: model_provider_name.into(),
+            raw_request: raw_request.clone(),
+            raw_response: None,
+            inference_params: InferenceParams::default(),
+            function_name: "",
+            variant_name: "",
+            dynamic_output_schema: None,
+            templates: &templates,
+            tool_config: None,
+            cached: false,
+            extra_body: Default::default(),
+            extra_headers: Default::default(),
+            fetch_and_encode_input_files_before_inference: false,
+        };
+
+        let result = collect_chunks(collect_chunks_args).await.unwrap();
+        let chat_result = match result {
+            InferenceResult::Chat(chat_result) => chat_result,
+            InferenceResult::Json(_) => panic!("Expected Chat inference response"),
+        };
+
+        assert_eq!(chat_result.content.len(), 2);
+        // Order should be preserved: text first, then tool call
+        match &chat_result.content[0] {
+            ContentBlockChatOutput::Text(text) => {
+                assert_eq!(
+                    text.text,
+                    "I'll help you with that. Let me search for information."
+                );
+            }
+            _ => panic!("Expected text block first"),
+        }
+        match &chat_result.content[1] {
+            ContentBlockChatOutput::ToolCall(tool_call) => {
+                assert_eq!(tool_call.raw_name, "search");
+                assert_eq!(tool_call.raw_arguments, r#"{"query": "weather today"}"#);
+                assert_eq!(tool_call.id, "tool_1");
+            }
+            _ => panic!("Expected tool call block second"),
+        }
+
+        // Test case 5: Tool call with empty name parts that should result in empty final name
+        let chunks_case5 = vec![InferenceResultChunk::Chat(ChatInferenceResultChunk {
+            content: vec![ContentBlockChunk::ToolCall(ToolCallChunk {
+                id: "tool_1".to_string(),
+                raw_name: None,
+                raw_arguments: "{\"test\": true}".to_string(),
+            })],
+            created,
+            usage: Some(Usage {
+                input_tokens: 5,
+                output_tokens: 5,
+            }),
+            raw_response: "chunk1".to_string(),
+            latency,
+            finish_reason: Some(FinishReason::ToolCall),
+        })];
+
+        let collect_chunks_args = CollectChunksArgs {
+            inference_id,
+            episode_id,
+            value: chunks_case5,
+            system: None,
+            input_messages: vec![],
+            function: function_config.clone(),
+            model_name: model_name.into(),
+            model_provider_name: model_provider_name.into(),
+            raw_request: raw_request.clone(),
+            raw_response: None,
+            inference_params: InferenceParams::default(),
+            function_name: "",
+            variant_name: "",
+            dynamic_output_schema: None,
+            templates: &templates,
+            tool_config: None,
+            cached: false,
+            extra_body: Default::default(),
+            extra_headers: Default::default(),
+            fetch_and_encode_input_files_before_inference: false,
+        };
+
+        let result = collect_chunks(collect_chunks_args).await.unwrap();
+        let chat_result = match result {
+            InferenceResult::Chat(chat_result) => chat_result,
+            InferenceResult::Json(_) => panic!("Expected Chat inference response"),
+        };
+
+        assert_eq!(chat_result.content.len(), 1);
+        match &chat_result.content[0] {
+            ContentBlockChatOutput::ToolCall(tool_call) => {
+                assert_eq!(tool_call.raw_name, ""); // Should be empty string when no name provided
+                assert_eq!(tool_call.raw_arguments, r#"{"test": true}"#);
+                assert_eq!(tool_call.id, "tool_1");
+            }
+            _ => panic!("Expected tool call block"),
+        }
+
+        // Test case 6: Complex multi-tool name accumulation across multiple chunks
+        let chunks_case6 = vec![
+            InferenceResultChunk::Chat(ChatInferenceResultChunk {
+                content: vec![
+                    ContentBlockChunk::ToolCall(ToolCallChunk {
+                        id: "tool_1".to_string(),
+                        raw_name: Some("get_".to_string()),
+                        raw_arguments: "{\"lo".to_string(),
+                    }),
+                    ContentBlockChunk::ToolCall(ToolCallChunk {
+                        id: "tool_2".to_string(),
+                        raw_name: Some("cal".to_string()),
+                        raw_arguments: "{\"op".to_string(),
+                    }),
+                    ContentBlockChunk::ToolCall(ToolCallChunk {
+                        id: "tool_3".to_string(),
+                        raw_name: Some("send_".to_string()),
+                        raw_arguments: "{\"me".to_string(),
+                    }),
+                ],
+                created,
+                usage: None,
+                raw_response: "chunk1".to_string(),
+                latency,
+                finish_reason: None,
+            }),
+            InferenceResultChunk::Chat(ChatInferenceResultChunk {
+                content: vec![
+                    ContentBlockChunk::ToolCall(ToolCallChunk {
+                        id: "tool_1".to_string(),
+                        raw_name: Some("wea".to_string()),
+                        raw_arguments: "cation\": ".to_string(),
+                    }),
+                    ContentBlockChunk::ToolCall(ToolCallChunk {
+                        id: "tool_2".to_string(),
+                        raw_name: Some("cul".to_string()),
+                        raw_arguments: "eration\": ".to_string(),
+                    }),
+                    ContentBlockChunk::ToolCall(ToolCallChunk {
+                        id: "tool_3".to_string(),
+                        raw_name: Some("email".to_string()),
+                        raw_arguments: "ssage\": ".to_string(),
+                    }),
+                ],
+                created,
+                usage: None,
+                raw_response: "chunk2".to_string(),
+                latency,
+                finish_reason: None,
+            }),
+            InferenceResultChunk::Chat(ChatInferenceResultChunk {
+                content: vec![
+                    ContentBlockChunk::ToolCall(ToolCallChunk {
+                        id: "tool_1".to_string(),
+                        raw_name: Some("ther".to_string()),
+                        raw_arguments: "\"Paris\"}".to_string(),
+                    }),
+                    ContentBlockChunk::ToolCall(ToolCallChunk {
+                        id: "tool_2".to_string(),
+                        raw_name: Some("ate".to_string()),
+                        raw_arguments: "\"5*5\"}".to_string(),
+                    }),
+                    ContentBlockChunk::ToolCall(ToolCallChunk {
+                        id: "tool_3".to_string(),
+                        raw_name: None, // No more name parts
+                        raw_arguments: "\"Hello world\"}".to_string(),
+                    }),
+                ],
+                created,
+                usage: Some(Usage {
+                    input_tokens: 20,
+                    output_tokens: 30,
+                }),
+                raw_response: "chunk3".to_string(),
+                latency: Duration::from_millis(250),
+                finish_reason: Some(FinishReason::ToolCall),
+            }),
+        ];
+
+        let collect_chunks_args = CollectChunksArgs {
+            inference_id,
+            episode_id,
+            value: chunks_case6,
+            system: None,
+            input_messages: vec![],
+            function: function_config.clone(),
+            model_name: model_name.into(),
+            model_provider_name: model_provider_name.into(),
+            raw_request: raw_request.clone(),
+            raw_response: None,
+            inference_params: InferenceParams::default(),
+            function_name: "",
+            variant_name: "",
+            dynamic_output_schema: None,
+            templates: &templates,
+            tool_config: None,
+            cached: false,
+            extra_body: Default::default(),
+            extra_headers: Default::default(),
+            fetch_and_encode_input_files_before_inference: false,
+        };
+
+        let result = collect_chunks(collect_chunks_args).await.unwrap();
+        let chat_result = match result {
+            InferenceResult::Chat(chat_result) => chat_result,
+            InferenceResult::Json(_) => panic!("Expected Chat inference response"),
+        };
+
+        assert_eq!(chat_result.content.len(), 3);
+        match &chat_result.content[0] {
+            ContentBlockChatOutput::ToolCall(tool_call) => {
+                assert_eq!(tool_call.raw_name, "get_weather"); // "get_" + "wea" + "ther"
+                assert_eq!(tool_call.raw_arguments, r#"{"location": "Paris"}"#);
+                assert_eq!(tool_call.id, "tool_1");
+            }
+            _ => panic!("Expected first tool call block"),
+        }
+        match &chat_result.content[1] {
+            ContentBlockChatOutput::ToolCall(tool_call) => {
+                assert_eq!(tool_call.raw_name, "calculate"); // "cal" + "cul" + "ate"
+                assert_eq!(tool_call.raw_arguments, r#"{"operation": "5*5"}"#);
+                assert_eq!(tool_call.id, "tool_2");
+            }
+            _ => panic!("Expected second tool call block"),
+        }
+        match &chat_result.content[2] {
+            ContentBlockChatOutput::ToolCall(tool_call) => {
+                assert_eq!(tool_call.raw_name, "send_email"); // "send_" + "email"
+                assert_eq!(tool_call.raw_arguments, r#"{"message": "Hello world"}"#);
+                assert_eq!(tool_call.id, "tool_3");
+            }
+            _ => panic!("Expected third tool call block"),
+        }
+    }
+
+    #[test]
+    fn test_json_inference_result_chunk_from_provider_chunk() {
+        use std::time::Duration;
+
+        // Test case for ToolCall content
+        let tool_chunk = ProviderInferenceResponseChunk {
+            content: vec![ContentBlockChunk::ToolCall(ToolCallChunk {
+                id: "123".to_string(),
+                raw_arguments: "{\"key\": \"value\"}".to_string(),
+                raw_name: Some("test_tool".to_string()),
+            })],
+            created: 1234567890,
+            usage: Some(Usage {
+                input_tokens: 10,
+                output_tokens: 20,
+            }),
+            raw_response: "raw response".to_string(),
+            latency: Duration::from_secs(1),
+            finish_reason: Some(FinishReason::ToolCall),
+        };
+
+        let result = JsonInferenceResultChunk::from(tool_chunk);
+        assert_eq!(result.raw, Some("{\"key\": \"value\"}".to_string()));
+        assert_eq!(result.thought, None);
+        assert_eq!(result.created, 1234567890);
+        assert_eq!(result.raw_response, "raw response");
+        assert_eq!(result.latency, Duration::from_secs(1));
+        assert_eq!(
+            result.usage,
+            Some(Usage {
+                input_tokens: 10,
+                output_tokens: 20
+            })
+        );
+        assert_eq!(result.finish_reason, Some(FinishReason::ToolCall));
+        // Test case for Text content
+        let text_chunk = ProviderInferenceResponseChunk {
+            content: vec![ContentBlockChunk::Text(TextChunk {
+                id: "123".to_string(),
+                text: "some text".to_string(),
+            })],
+            created: 1234567890,
+            usage: None,
+            raw_response: "raw response".to_string(),
+            latency: Duration::from_secs(1),
+            finish_reason: None,
+        };
+
+        let result = JsonInferenceResultChunk::from(text_chunk);
+        assert_eq!(result.raw, Some("some text".to_string()));
+        assert_eq!(result.thought, None);
+
+        // Test case for Thought content
+        let thought_chunk = ProviderInferenceResponseChunk {
+            content: vec![ContentBlockChunk::Thought(ThoughtChunk {
+                id: "123".to_string(),
+                text: Some("thinking...".to_string()),
+                signature: None,
+                provider_type: None,
+            })],
+            created: 1234567890,
+            usage: None,
+            raw_response: "raw response".to_string(),
+            latency: Duration::from_secs(1),
+            finish_reason: None,
+        };
+
+        let result = JsonInferenceResultChunk::from(thought_chunk);
+        assert_eq!(result.raw, None);
+        assert_eq!(result.thought, Some("thinking...".to_string()));
+        assert_eq!(result.finish_reason, None);
+        // Test case for multiple content blocks - should use last raw content
+        let mixed_chunk = ProviderInferenceResponseChunk {
+            content: vec![
+                ContentBlockChunk::Text(TextChunk {
+                    id: "123".to_string(),
+                    text: "first text".to_string(),
+                }),
+                ContentBlockChunk::ToolCall(ToolCallChunk {
+                    id: "456".to_string(),
+                    raw_arguments: "final content".to_string(),
+                    raw_name: Some("test_tool".to_string()),
+                }),
+                ContentBlockChunk::Thought(ThoughtChunk {
+                    id: "789".to_string(),
+                    text: Some("final thought".to_string()),
+                    signature: None,
+                    provider_type: None,
+                }),
+            ],
+            created: 1234567890,
+            usage: None,
+            raw_response: "raw response".to_string(),
+            latency: Duration::from_secs(1),
+            finish_reason: None,
+        };
+
+        let result = JsonInferenceResultChunk::from(mixed_chunk);
+        assert_eq!(result.raw, Some("final content".to_string()));
+        assert_eq!(result.thought, Some("final thought".to_string()));
+
+        // Test case for empty content
+        let empty_chunk = ProviderInferenceResponseChunk {
+            content: vec![],
+            created: 1234567890,
+            usage: None,
+            raw_response: "raw response".to_string(),
+            latency: Duration::from_secs(1),
+            finish_reason: None,
+        };
+
+        let result = JsonInferenceResultChunk::from(empty_chunk);
+        assert_eq!(result.raw, None);
+        assert_eq!(result.thought, None);
+        assert_eq!(result.finish_reason, None);
     }
 }
